Modified submission test data for SubmissionSpec to live in spec rather than OrientDbTestFixture, reused initialization code for consistency and to ensure PoisonPill even if one test fails
